Replace the tickadj dithering with a better, simpler algorithm
Instead of the crazy thing that I invented before, just use an
accumulator to track "sub-tick" resolution and add an extra tick
whenever the accumulator overflows. This gives better absolute phase
noise (because the extra ticks are more regularly spaced) and better
tracking (because the accumulator is preserved when tickadj changes and
this does the right thing, especially in the face of +/- 1 unit changes
like you'd see in the steady state).

Inspiration: TI technical brief SWRA029 "Fractional/Integer-N PLL
Basics".